dia-stack
pass without lambdaS3Policy
